Upgrade to Bevy 0.17, follow migration guide to refactor errors (#107)
Most of the changes are related to renames of various components
exported by Bevy.

This will need https://github.com/vleue/polyanya/pull/119 to land. I
have assumed when that does land that Polyanya will be updated to
`0.15.0` and so consequently I've gone ahead and guessed that version in
Cargo.toml (I tested the changes with `path = ../polyanya`, while
testing that PR to ensure everything was passing here.

---------

Co-authored-by: François Mockers <francois.mockers@vleue.com>